requirements.txt
pandas
numpy
faiss-cpu
sentence-transformers
torch
requests
streamlit

